# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: RASHED — AI Water Intelligence API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateMe
      tags: [auth]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"

  # ─── HOUSEHOLD ───
...
  /analysis/latest:
    get:
      operationId: getLatestAnalysis
      tags: [analysis]
      summary: Get the latest consumption analysis
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsumptionAnalysis"
        "404":
          description: No analysis yet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /consumption/history:
    get:
      operationId: getConsumptionHistory
      tags: [consumption]
      summary: Get consumption history
      parameters:
        - name: months
          in: query
          schema:
            type: number
            default: 6
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConsumptionRecord"
...
  /forecast:
    get:
      operationId: getForecast
      tags: [forecast]
      summary: Get current month consumption forecast
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Forecast"
        "404":
          description: Not enough data for forecast
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── RECOMMENDATIONS ───
...
  /simulation/what-if:
    post:
      operationId: runWhatIfSimulation
      tags: [simulation]
      summary: Run a what-if simulation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WhatIfInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WhatIfResult"

  # ─── REFERENCE DATA ───
...
  /assistant/conversations:
    get:
      operationId: listAssistantConversations
      tags: [assistant]
      summary: List AI assistant conversations
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AssistantConversation"
    post:
      operationId: createAssistantConversation
      tags: [assistant]
      summary: Create a new assistant conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssistantConversationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssistantConversation"
...
  /assistant/conversations/{id}/messages:
    get:
      operationId: listAssistantMessages
      tags: [assistant]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AssistantMessage"
    post:
      operationId: sendAssistantMessage
      tags: [assistant]
      summary: Send a message to the assistant (SSE stream with function calling)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssistantMessageInput"
      responses:
        "200":
          description: SSE stream of assistant response
          content:
            text/event-stream: {}

  # ─── GEMINI (from skill) ───
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    RegisterInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        preferredLanguage:
          type: string
          enum: [ar, en]
          default: ar
      required:
        - name
        - email
        - password
    LoginInput:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    AuthUser:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        email:
          type: string
        preferredLanguage:
          type: string
        householdId:
          type: ["number", "null"]
        isDemoMode:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - preferredLanguage
        - isDemoMode
        - createdAt
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        preferredLanguage:
          type: string
          enum: [ar, en]

    # ─── HOUSEHOLD ───
...
    ConsumptionAnalysis:
      type: object
      properties:
        id:
          type: number
        billId:
          type: number
        currentConsumptionM3:
          type: ["number", "null"]
        previousConsumptionM3:
          type: ["number", "null"]
        changePercent:
          type: ["number", "null"]
        anomalyDetected:
          type: boolean
        anomalyExplanation:
          type: ["string", "null"]
        leakRiskLevel:
          type: ["string", "null"]
          enum: [low, medium, high, null]
        leakRiskProbability:
          type: ["number", "null"]
        leakRiskReason:
          type: ["string", "null"]
        whyConsumptionIncreased:
          type: array
          items:
            $ref: "#/components/schemas/ConsumptionCause"
        dataClassification:
          type: string
        tariffConfigVersion:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - billId
        - anomalyDetected
        - whyConsumptionIncreased
        - dataClassification
        - createdAt
    ConsumptionCause:
      type: object
      properties:
        cause:
          type: string
        confidence:
          type: string
          enum: [low, medium, high]
        explanation:
          type: string
      required:
        - cause
        - confidence
        - explanation

    # ─── CONSUMPTION ───
    ConsumptionRecord:
      type: object
      properties:
        id:
          type: number
        householdId:
          type: number
        periodStart:
          type: ["string", "null"]
        periodEnd:
          type: ["string", "null"]
        consumptionM3:
          type: number
        estimatedCostSar:
          type: ["number", "null"]
        tariffConfigVersion:
          type: ["string", "null"]
        dataClassification:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - householdId
        - consumptionM3
        - dataClassification
        - createdAt
...
    Forecast:
      type: object
      properties:
        id:
          type: number
        householdId:
          type: number
        currentConsumptionM3:
          type: number
        projectedEndOfMonthM3:
          type: number
        expectedRangeMinM3:
          type: number
        expectedRangeMaxM3:
          type: number
        daysElapsed:
          type: number
        daysInMonth:
          type: number
        confidenceLevel:
          type: string
          enum: [low, medium, high]
        confidenceNote:
          type: ["string", "null"]
        estimatedCostSar:
          type: ["number", "null"]
        tariffConfigVersion:
          type: ["string", "null"]
        dataClassification:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - householdId
        - currentConsumptionM3
        - projectedEndOfMonthM3
        - expectedRangeMinM3
        - expectedRangeMaxM3
        - daysElapsed
        - daysInMonth
        - confidenceLevel
        - dataClassification
        - createdAt

    # ─── RECOMMENDATIONS ───
...
    WhatIfInput:
      type: object
      properties:
        reductionPercent:
          type: number
          minimum: 1
          maximum: 100
      required:
        - reductionPercent
    WhatIfResult:
      type: object
      properties:
        currentConsumptionM3:
          type: number
        projectedConsumptionM3:
          type: number
        deltaM3:
          type: number
        currentCostSar:
          type: number
        projectedCostSar:
          type: number
        deltaSar:
          type: number
        reductionPercent:
          type: number
        tariffConfigVersion:
          type: string
        verificationStatus:
          type: string
        dataClassification:
          type: string
      required:
        - currentConsumptionM3
        - projectedConsumptionM3
        - deltaM3
        - currentCostSar
        - projectedCostSar
        - deltaSar
        - reductionPercent
        - tariffConfigVersion
        - verificationStatus
        - dataClassification

    # ─── REFERENCE ───
...
    AssistantConversation:
      type: object
      properties:
        id:
          type: number
        userId:
          type: number
        householdId:
          type: ["number", "null"]
        title:
          type: string
        language:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - title
        - language
        - createdAt
    AssistantConversationInput:
      type: object
      properties:
        title:
          type: string
        language:
          type: string
          enum: [ar, en]
      required:
        - title
...
    AssistantMessage:
      type: object
      properties:
        id:
          type: number
        conversationId:
          type: number
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        toolCallsSummary:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt
    AssistantMessageInput:
      type: object
      properties:
        content:
          type: string
        language:
          type: string
          enum: [ar, en]
      required:
        - content

    # ─── GEMINI (from skill) ───
...